linked countdown to image capture, decreased counter minimum,
hooked up jitter portion of the patch (when the image is grabbed) to
the metro, so that an image is grabbed when the counter reaches 0
(visually, the monitor will read ‘smile!’ instead of a 0 displaying on
the screen).

altered the counter minimum to 0, it was previously 1.

bug that needs fixing: counter + carrybang — sending out bangs when the
counter reaches the minimum and maximum value. i only want a bang sent
when the minimum value is reached. how to fix this? after reading some
posts on a max msp forum, i think that this might mean that i need to
manipulate the outlets/inlets or coordinate them with the counter
because its counting down rather than up? idk.