export image issue (recurring)
hooked up ‘loadbang’ to the button triggered by the counter to make
sure that jit.grab is ready to get to work. export image is also
connected to the button triggered by the counter, but when i run the
system, no image is exported. does this have something to do with the
patch being saved in a github folder? i dont think so, but maybe. im
going to save the patch somewhere else on my cpu and see if the problem
is fixed. fingers crossed :)