exportimage bug fixed B)
Had to simply close and reopen the patch to get the image capture (once
the counter hits 0) running properly. I moved the patch to a folder on
my desktop, then closed the patch, then open and ran it, and it worked.
I then replaced the older patch (saved in our shared repo, the one i’m
pushing now) with this one.

katherine’s explanation: the reference to the file wasn’t mapping
right, with regard to which file you are currently running. hope that
makes sense